requirements.txt
streamlit
langchain
langchain-openai
langchain_community
langchain-text-splitters
pypdf
icecream
python-dotenv
faiss-cpu